refactor: move memory and time measurements to utils
